+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2
+
+- Fix foreign key definitions not being in the end.
+- Add foreign key test group with one test.
+
 ## 0.1.1
 
 - Fix bug where I have assumed that foreign key fields are automatically NOT NULL.